use `size_t` as the type used for `isc_counter_t`
`isc_counter_t` was using a plethora of integer types internally and in
it's API. Change it to be consistent and use `size_t` only.

+++ lib/isc/include/isc/counter.h
@@ -39,7 +39,7 @@
 *****/
 
 void
-isc_counter_create(isc_mem_t *mctx, int limit, isc_counter_t **counterp);
+isc_counter_create(isc_mem_t *mctx, size_t limit, isc_counter_t **counterp);
 /*%<
  * Allocate and initialize a counter object.
  */
@@ -54,19 +54,19 @@ isc_counter_increment(isc_counter_t *counter);
  * incremented regardless of return value.)
  */
 
-unsigned int
+size_t
 isc_counter_used(isc_counter_t *counter);
 /*%<
  * Return the current counter value.
  */
 
 void
-isc_counter_setlimit(isc_counter_t *counter, int limit);
+isc_counter_setlimit(isc_counter_t *counter, size_t limit);
 /*%<
  * Set the counter limit.
  */
 
-unsigned int
+size_t
 isc_counter_getlimit(isc_counter_t *counter);
 /*%<
  * Get the counter limit.
